Move ensure_datetime to utils and fix circular import
